masih error pas update gambar di kategori materi

+++ app/Http/Controllers/MateriController.php
@@ -56,20 +56,23 @@ class MateriController extends Controller
     public function updateCategory(Request $request)
     {
         $category = Category::find($request->input('id'));
-
         $category->name = $request->input('name');
         $category->save();
+        
+        if ($request->base64_image) {
+            $image = $category->image;
+            if ($image) {
+                Storage::delete($image->filename);
+            }
+            $newImg = base64_decode($request->base64_image);
+            $str = Str::random(10);
+            $filename = 'public/category/' . (string) $category->id . $request->input('name') . '$' . $str . '.jpg';
+            Storage::put($filename, $newImg);
+            $image->filename = $filename;
+            $image->save();
+        }
 
-        $image = $category->image;
-        Storage::delete($image->filename);
-        $newImg = base64_decode($request->input('base64_img'));
-        $str = Str::random(10);
-        $filename = 'public/category/' . (string) $category->id . $request->input('name') . '$' . $str . '.jpg';
-        Storage::put($filename, $newImg);
-
-        $image->filename = $filename;
-        $image->save();
-
+     
         return response()->json([
             'message' => 'category is updated'
         ]);
@@ -228,8 +231,8 @@ class MateriController extends Controller
     {
         return MateriRes::collection(
             Materi::where('title', 'LIKE', '%' . $request->keyword . '%')
-            ->orWhere('content', 'LIKE', '%' . $request->keyword . '%')
-            ->get()
+                ->orWhere('content', 'LIKE', '%' . $request->keyword . '%')
+                ->get()
         );
     }
